% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_basics_package.R
\name{add_biz_column}
\alias{add_biz_column}
\title{Adds a biz id column but keeps bid}
\usage{
add_biz_column(df, bizbid_table)
}
\arguments{
\item{bizbid_table}{dataframe with 2 cols bid and bizid, can use get_bizid_map() function}

\item{redshift}{redshift object}
}
\description{
This function collects cheque data from a googlesheet, cleans and returns a consistent dataframe for analysis.
}
\examples{
add_biz_column(myRedshift, bizbid_table)


}
\keyword{conversion}
\keyword{utility}
