% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_basics_package.R
\name{get_bizbid_map}
\alias{get_bizbid_map}
\title{Gets the bizid bid map}
\usage{
get_bizbid_map(myRedshift)
}
\description{
This function collects cheque data from a googlesheet, cleans and returns a consistent dataframe for analysis.
}
\examples{
get_bizbid_map(myRedshift)


}
\keyword{object}
\keyword{redshift}
\keyword{sql}
\keyword{utility}
