% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_basics_package.R
\name{get_credit_customers}
\alias{get_credit_customers}
\title{Gets credit customer data}
\usage{
get_credit_customers(myRedshift)
}
\arguments{
\item{redshift}{redshift object}
}
\description{
This function collects cheque data from a googlesheet, cleans and returns a consistent dataframe for analysis.
}
\examples{
get_credit_customers(myRedshift)


}
\keyword{sql}
\keyword{utility}
