% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_basics_package.R
\name{get_customer_transactions}
\alias{get_customer_transactions}
\title{Takes Payments Data and gives transactions by bizid and date}
\usage{
get_customer_transactions(paymentsdata)
}
\arguments{
\item{payments}{redshift object}
}
\description{
This function collects cheque data from a googlesheet, cleans and returns a consistent dataframe for analysis.
}
\examples{
get_customer_transactions(paymentsdata)

}
\keyword{daily}
\keyword{transactions}
