% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_basics_package.R
\name{get_jt_storenamedata}
\alias{get_jt_storenamedata}
\title{Gets store names data}
\usage{
get_jt_storenamedata(myRedshift)
}
\arguments{
\item{redshift}{redshift object}
}
\description{
This function collects cheque data from a googlesheet, cleans and returns a consistent dataframe for analysis.
}
\examples{
get_jt_storenamedata(myRedshift)


}
\keyword{sql}
\keyword{utility}
