% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_basics_package.R
\name{get_payments_data}
\alias{get_payments_data}
\title{Wrapper around paymetns data that only loads paymentsdata if it's not already up to date 
Load Payments Data since Jan to most recent available}
\usage{
get_payments_data()
}
\arguments{
\item{none}{no parameters}
}
\description{
This function collects payments data from across several googlesheets, cleans and produces a consistent dataframe for analysis.
}
\examples{
get_payments_data()

}
\keyword{data}
\keyword{googlesheets}
\keyword{payments}
