% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_basics_package.R
\name{get_paymentsdata_sqlite}
\alias{get_paymentsdata_sqlite}
\title{Load Payments Data since Jan to most recent available}
\usage{
get_paymentsdata_sqlite(dbname)
}
\arguments{
\item{none}{no parameters}
}
\description{
This function collects payments data from across several googlesheets, cleans and produces a consistent dataframe for analysis.
}
\examples{
load_payments_data()

Another payment data extraction function

This function pulls paymentsdata from a local sqlite database
db <- get_paymentsdata_sqlite(dbname)
}
\keyword{data}
\keyword{googlesheets}
\keyword{helper}
\keyword{payments}
