% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_basics_package.R
\name{getmonthlysheet}
\alias{getmonthlysheet}
\title{A payment data extraction function}
\usage{
getmonthlysheet(filenamevar)
}
\arguments{
\item{filename}{String filename of the gsheet that holds the data}
}
\description{
This function is a helper function for the load_payments_data function below
}
\examples{
db <- getmonthlysheet("filename")
}
\keyword{helper}
