% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_basics_package.R
\name{load_cheques_data}
\alias{load_cheques_data}
\title{Load Cheque Data since Jan to most recent available}
\usage{
load_cheques_data()
}
\arguments{
\item{none}{no parameters}
}
\description{
This function collects cheque data from a googlesheet, cleans and returns a consistent dataframe for analysis.
}
\examples{
load_cheques_data()


}
\keyword{cheques}
\keyword{data}
\keyword{googlesheets}
\keyword{payments}
