% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_basics_package.R
\name{load_orders_data}
\alias{load_orders_data}
\title{Load Orders Data since April 16, 2016 to recent available}
\usage{
load_orders_data(con1)
}
\arguments{
\item{jthost}{host url for Amazon Redshift Database}

\item{jtport}{port for Amazon Redshift Database}

\item{jtdbname}{database name for Amazon Redshift Database}

\item{redshift_user}{password for Amazon Redshift Database}

\item{redshift_password}{password for Amazon Redshift Database}
}
\description{
This function collects cheque data from a googlesheet, cleans and returns a consistent dataframe for analysis.
}
\examples{
load_orders_data(databaseconnection)


}
\keyword{cheques}
\keyword{data}
\keyword{googlesheets}
\keyword{payments}
