% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_basics_package.R
\name{load_payments_data}
\alias{load_payments_data}
\title{Payment data extraction & cleaning function}
\usage{
load_payments_data()
}
\description{
This function visits the monthly reconciliation sheets and collects the data together
}
\examples{
db <- load_payments_data() )
}
\keyword{payments}
\keyword{sqlite}
