% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_basics_package.R
\name{swap_month_day}
\alias{swap_month_day}
\title{A date utility function}
\usage{
swap_month_day(inpdate)
}
\arguments{
\item{inpdate}{String date}
}
\description{
This function is a helper function that swaps months and days of a date
}
\examples{
swap_month_day(inpdate)
}
\keyword{helper}
