% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_basics_package.R
\name{update_pmtsdata_sqlite}
\alias{update_pmtsdata_sqlite}
\title{Another payment data extraction function}
\usage{
update_pmtsdata_sqlite(filename, inpdate)
}
\arguments{
\item{filename}{String filename of the gsheet that holds the new data}

\item{inpdate}{String date after which data is considered new}
}
\description{
This function collects new payments data and updates the local sqlite database
}
\examples{
db <- update_pmtsdata_sqlite("filename", inpdate = max(paymentsdata$deliverydate, na.rm=T) )
}
\keyword{payments}
\keyword{sqlite}
