% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_basics_package.R
\name{write_to_gs}
\alias{write_to_gs}
\title{Writes Function to Googlesheet}
\usage{
write_to_gs(df, df_name)
}
\arguments{
\item{df}{dataframe to write}

\item{df_name}{name of file to write in Googlesheets}
}
\description{
This function collects cheque data from a googlesheet, cleans and returns a consistent dataframe for analysis.
}
\examples{
write_to_gs(df, df_name)


}
\keyword{googlesheets}
\keyword{utility}
